Increase tolerance for offer
If power users who have 100-200 offers go online/offline the
offers in the network gets very volatile for a while.
Currently the publishing is streamed over quite a long delay
and at shutdown not all offer might be removed so TTL will
trigger removal later, so it is expected that nodes have
different view on offers. We work on improvements to avoid
that volatility, but unitl that is achieved we increase
tolerance for offers.