Add support for isApiUser and check in OpenOfferManager
Calls from the API set the flag to true, from the desktop its false.
If offer has denyApiUser enabled the offer availibility check fails.

@ghubstan: I set the value in CoreApi and not directly in the CoreTradesService,
so keep it more open that those sub domain classes can be used from desktop as
well. Not sure if is not intended/permitted for some reason but I though that
way we have more flexibility. Feel free to remove the param and set it directly
in CoreTradesService if that class will never be used from non-api code.