Use posix-style method options in CLI
All api methods that take one or more parameters now use posix-style params
instead of positional params.  This makes method params less ambiguous,
and provides some client side validation that was absent before this change.

This is ongoing work, as the CLI level help text needs to be adjusted for
the new parameter style, and server side method help needs to be written.