Refactor apitest's bot related classes
A new botsupport package was added, with no dependencies on the test
harness.  The intent is eventually move this botsupport pkg code to a
new gradle subproject called ':bot' which would depend only on the
':cli' subproject (to keep the ':cli' module light).

The future ':bot' subproject would not be shipped, but would give java
devs some api java-bot support and at least one example bot.  A bsq-btc
market maker bot is included in this commit.