Workaround API editoffer event rcv order problem in peer UI
When the API is used to edit an offer, the daemon's P2PDataStorage will
send onRemoved(offer) and onAdded(offer) events 1-3 ms apart, and those
events may arrive at a peer's UI in the wrong order: onAdded, onRemoved.
When the order is backwards, the edited offer disappears from the UI's
OfferBook view.

Reloading the offerBookListItems works around the problem.  Reloading
100 offers usually takes less then 1 ms.

I would prefer to solve this in 2+ other classes:

- P2PDataStorage, where these events are fired in the correct order,
  but down to 1ms apart.

- OfferBookService constructor, where the peer UI's HashMapChangedListener
  methods 'onRemoved' and 'onAdded' are causing API edited offers to
  dissappear if received in backwards order:  'onAdded', then 'onRemoved'.

Trying to insert a delay between remove an offer for edit, then
republishing the offer, does not work.  The events can still
arrive in the wrong order.

I am going to need some help from @chimp1984 to solve it.