Conditionally display altcoin offer trigger price
Refactored some CLI console output formatting code to adjust to new
XMR offer support in API.  (TableFormat's 'formatOfferTable' method
was moved to a new OfferFormat class.)

BSQ offers cannot have a margin-based-price, and trigger-price was
never displayed.  XMR offers can have a margin-based-price and a
trigger-price, so altcoin offer display in the console is not as
simple as before.