Correctly scale CLI's 'editoffer' trigger-price param
Unfortunately, the EditOffer proto triggerPrice field was declared as
a long instead of a string, so the CLI has to look at the offer to know
how to scale the trigger-price (for a fiat or altcoin offer) param sent
to the server in its 'editoffer' request.  That means a preliminary round
trip to the server:  a 'getmyoffer' request.