When daoActivated is set to false the DAO P2P data are still received from seed nodes and processed but as the services for processing the payloads are not added the data is inefficiently processed. The getMap returned a flattened map of all maps in all services which can be quite large. We use now a filtered map with calling canHandle first. Also the put got optimized to indicate in the return value if there has been a service found to add the payload. If not we do not invoke the listeners and do not broadcast.
To not request the DAO P2P data would be better but I
don't see a easy way how to do that as the P2P network
is not aware of the type of data. Some market interface
could be used and a flag at the request to the seed node
to indicate if those types should be included but that
does feel too customized for a special use case. The
DAO P2P data is not that big as well, so I think for now
that fix should be good enough.