Move volume formatting from DisplayUtils to VolumeUtil
In practice, presentation/display logic should be kept in UIs, but the core
api will need to duplicate a very small part of the code in DisplayUtils if
a very small part of this refactoring is not done.  Instead of moving a tiny
bit of the volume formatting methods, all of them are moved to core.