Refactor desktop Closed Tradable model utils for API use
Move some potentially sharable utility code from desktop's ClosedTrade
models to core's new ClosedTradeUtil.

The API's `gettrades --category=<open|closed|failed>` method will need
some of this logic on the server side, but the scope of this refactoring
is larger in the interest of a more complete refactor, and avoiding
duplication in the future.