Add scriptTypeId to RawTransactionInput
We will need that to know if its a segwit input or not.
Add constructors for converting a TransactionInput to
RawTransactionInput. Handle bitcoinSerialize according
to segwit flag.

This change should not break anything as we only add a
field to protobuf and RawTransactionInput is not used in
cases where we depend on the hash of the payload like in
offers.