Reset address entry if BSQ swap offer is used for create offer
We create a new address entry when opening the create offer view.
But we close and replace that view with the BSQ swap create offer view if
BSQ swap is selected, therefore the address entry would be dangling and
later cleaned up at openOfferManager. To clean it up directly we call the
resetAddressEntriesForOpenOffer on btcWalletService.