Remove prevHash from StateHash classes.
We used prevHash only for display in the UI but it did
not had any functional usage. It took about 50% of the data size
of the StateHashes (about 3 MB out of 6 MB total storage size).
The prevHash is used only as part of the pre image for the hash
but we take the hash of the previous StateHash item and then
stored that in the newly created StateHash item. So we do not change
that the previous item is part of the new hash, but only that we do not
persist it as it is not required and helps reduce the data load.